BACKEND/requirements.txt
asgiref==3.2.8
beautifulsoup4==4.9.1
certifi==2020.4.5.2
chardet==3.0.4
cmake==3.18.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.1.15
django-allauth==0.42.0
django-bootstrap-pagination==1.7.1
django-cors-headers==3.3.0
django-extensions==2.2.9
django-heroku==0.3.1
django-random-queryset==0.1.3
django-rest-auth==0.9.5
django-rest-passwordreset==1.1.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
dlib==19.20.0
drf-yasg==1.17.1
Faker==4.1.0
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.1
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msgpack-python==0.5.6
msgpack-rpc-python==0.4.1
numpy==1.18.5
oauthlib==3.1.0
openapi-codec==1.3.2
opencv-contrib-python==4.3.0.36
packaging==20.4
psycopg2==2.8.5
psycopg2-binary==2.8.5
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
text-unidecode==1.3
tornado==4.5.3
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.1.0
zipp==1.0.0
